refactor: onboarding steps의 step3, step4 폴더 구조 및 네이밍 표준화
