+++ src/pages/onboarding/components/steps/activityInfo/ActivityInfo.tsx
