+++ src/pages/onboarding/components/steps/interiorStyle/InteriorStyle.css.ts
