+++ src/pages/onboarding/components/steps/interiorStyle/MoodBoard.css.ts
