+++ src/pages/onboarding/components/steps/interiorStyle/MoodBoard.tsx
